Extract frame sync helpers
